Only use one thread for building

+++ engenious.ContentTool/Builder/ContentBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
@@ -11,7 +12,7 @@ using Mono.Cecil;
 
 namespace engenious.ContentTool.Builder
 {
-    public class ContentBuilder
+    public class ContentBuilder : IDisposable
     {
         public const string FileExtension = ".ego";
 
@@ -25,7 +26,33 @@ namespace engenious.ContentTool.Builder
 
         private readonly BuildCache _cache;
 
-        private Thread _buildThread;
+        private readonly Thread _buildThread;
+        private readonly ConcurrentQueue<BuildTask> _buildQueue;
+        private readonly AutoResetEvent _startBuild;
+        private readonly CancellationTokenSource _buildThreadCancellation;
+        private readonly CancellationTokenSource _buildCancellation;
+
+        public enum BuildTaskType
+        {
+            Clean,
+            Build,
+            Rebuild
+        }
+        public class BuildTask
+        {
+            public BuildTask(BuildTaskType buildTaskType, ContentItem buildItem = null)
+            {
+                BuildTaskType = buildTaskType;
+                BuildItem = buildItem;
+                CancellationTokenSource = new CancellationTokenSource();
+            }
+
+            public BuildTaskType BuildTaskType { get; }
+            
+            public ContentItem BuildItem { get; }
+            
+            public CancellationTokenSource CancellationTokenSource { get; }
+        }
 
         public ContentBuilder(ContentProject project, IRenderingSurface renderingSurface = null,
             GraphicsDevice graphicsDevice = null)
@@ -38,6 +65,48 @@ namespace engenious.ContentTool.Builder
             _cache = BuildCache.Load(Path.Combine(Path.GetDirectoryName(project.ContentProjectPath), "obj",
                 project.Configuration,
                 project.Name + ".dat"));
+            
+            
+            _buildThreadCancellation = new CancellationTokenSource();
+            var buildThreadToken = _buildThreadCancellation.Token;
+            buildThreadToken.Register(() => _startBuild.Set());
+            
+            _buildQueue = new ConcurrentQueue<BuildTask>();
+            
+            _buildThread = new Thread((cancellationObj) =>
+            {
+                var cancellationToken = (CancellationToken)(cancellationObj ?? throw new InvalidCastException());
+                while(!cancellationToken.IsCancellationRequested)
+                {
+                    _startBuild.WaitOne();
+                    if (cancellationToken.IsCancellationRequested)
+                        return;
+                    while (_buildQueue.TryDequeue(out var buildTask))
+                    {
+                        IsBuilding = true;
+                        switch (buildTask.BuildTaskType)
+                        {
+                            case BuildTaskType.Clean:
+                                CleanThread(buildTask.CancellationTokenSource.Token);
+                                break;
+                            case BuildTaskType.Build:
+                                BuildThread(buildTask.BuildItem, buildTask.CancellationTokenSource.Token);
+                                break;
+                            case BuildTaskType.Rebuild:
+                                if (Project.HasUnsavedChanges)
+                                    Project.Save();
+                                CleanThread(buildTask.CancellationTokenSource.Token);
+                                BuildThread(Project, buildTask.CancellationTokenSource.Token);
+                                break;
+                            default:
+                                throw new ArgumentOutOfRangeException();
+                        }
+                        IsBuilding = false;
+                    }
+                }
+            });
+
+            _buildThread.Start(buildThreadToken);
         }
 
         public void Build()
@@ -45,44 +114,28 @@ namespace engenious.ContentTool.Builder
             Build(Project);
         }
 
-        public void Build(ContentItem item)
+        private BuildTask EnqueueBuildTask(BuildTask task)
         {
-            _buildThread = new Thread(() =>
-            {
-                IsBuilding = true;
-                BuildThread(item);
-                IsBuilding = false;
-            });
-            //_buildThread.SetApartmentState(ApartmentState.STA);
-            _buildThread.Start();
+            _buildQueue.Enqueue(task);
+
+            _startBuild.Set();
+            
+            return task;
         }
 
-        public void Clean()
+        public BuildTask Build(ContentItem item)
         {
-            _buildThread = new Thread(() =>
-            {
-                IsBuilding = true;
-                CleanThread();
-                IsBuilding = false;
-            });
-            //_buildThread.SetApartmentState(ApartmentState.STA);
-            _buildThread.Start();
+            return EnqueueBuildTask(new BuildTask(BuildTaskType.Build, item));
         }
 
-        public void Rebuild()
+        public BuildTask Clean()
         {
-            _buildThread = new Thread(() =>
-            {
-                IsBuilding = true;
+            return EnqueueBuildTask(new BuildTask(BuildTaskType.Clean));
+        }
 
-                if (Project.HasUnsavedChanges)
-                    Project.Save();
-                CleanThread();
-                BuildThread(Project);
-                IsBuilding = false;
-            });
-            //_buildThread.SetApartmentState(ApartmentState.STA);
-            _buildThread.Start();
+        public BuildTask Rebuild()
+        {
+            return EnqueueBuildTask(new BuildTask(BuildTaskType.Rebuild));
         }
 
         public void Abort()
@@ -91,15 +144,10 @@ namespace engenious.ContentTool.Builder
                 _buildThread.Abort();
         }
 
-        public void Join()
-        {
-            _buildThread.Join();
-        }
-
         internal IRenderingSurface RenderingSurface { get; }
         internal GraphicsDevice GraphicsDevice { get; }
 
-        protected void BuildThread(ContentItem item)
+        protected void BuildThread(ContentItem item, CancellationToken cancellationToken)
         {
             FailedBuilds = 0;
             PipelineHelper.PreBuilt(Project);
@@ -127,7 +175,7 @@ namespace engenious.ContentTool.Builder
 
                 iContext.BuildMessage += RaiseBuildMessage;
                 pContext.BuildMessage += RaiseBuildMessage;
-                InternalBuildItem(item, outputDestination, iContext, pContext);
+                InternalBuildItem(item, outputDestination, iContext, pContext, cancellationToken);
                 try
                 {
                     if (rewriteAssembly)
@@ -146,10 +194,12 @@ namespace engenious.ContentTool.Builder
             _cache.Save();
         }
 
-        protected void CleanThread()
+        protected void CleanThread(CancellationToken cancellationToken)
         {
             foreach (var item in _cache.Files)
             {
+                if (cancellationToken.IsCancellationRequested)
+                    return;
                 if (File.Exists(item.Value.OutputFilePath))
                     File.Delete(item.Value.OutputFilePath);
             }
@@ -163,9 +213,12 @@ namespace engenious.ContentTool.Builder
             }
         }
 
-        protected void InternalBuildItem(ContentItem item, string outputDestination,
-            ContentImporterContext importerContext, ContentProcessorContext processorContext)
+        private void InternalBuildItem(ContentItem item, string outputDestination,
+            ContentImporterContext importerContext, ContentProcessorContext processorContext, CancellationToken cancellationToken)
         {
+            if (cancellationToken.IsCancellationRequested)
+                return;
+            
             if (!(item is ContentProject))
                 outputDestination = Path.Combine(outputDestination, item.Name);
 
@@ -173,7 +226,9 @@ namespace engenious.ContentTool.Builder
             if (folder != null)
             {
                 foreach (var child in folder.Content)
-                    InternalBuildItem(child, outputDestination, importerContext, processorContext);
+                {
+                    InternalBuildItem(child, outputDestination, importerContext, processorContext, cancellationToken);
+                }
             }
             else
             {
@@ -182,7 +237,7 @@ namespace engenious.ContentTool.Builder
                     InternalBuildFile(item as ContentFile,
                         Path.Combine(Path.GetDirectoryName(outputDestination),
                             Path.GetFileNameWithoutExtension(item.Name) + FileExtension), importerContext,
-                        processorContext);
+                        processorContext, cancellationToken);
                     RaiseBuildMessage(this,
                         new BuildMessageEventArgs(item.RelativePath, item.RelativePath + " built",
                             BuildMessageEventArgs.BuildMessageType.Information));
@@ -202,9 +257,9 @@ namespace engenious.ContentTool.Builder
         protected void RaiseBuildMessage(object sender, BuildMessageEventArgs e) => BuildMessage?.Invoke(e);
 
         public object InternalBuildFile(ContentFile item, string destination, ContentImporterContext importerContext,
-            ContentProcessorContext processorContext)
+            ContentProcessorContext processorContext, CancellationToken cancellationToken)
         {
-            return BuildFile(item, destination, importerContext, processorContext, _cache);
+            return BuildFile(item, destination, importerContext, processorContext, cancellationToken, _cache);
         }
 
         /// <summary>
@@ -216,19 +271,23 @@ namespace engenious.ContentTool.Builder
         /// <param name="processorContext"></param>
         /// <returns></returns>
         public static object BuildFile(ContentFile item, string destination, ContentImporterContext importerContext,
-            ContentProcessorContext processorContext, BuildCache cache = null)
+            ContentProcessorContext processorContext, CancellationToken cancellationToken, BuildCache cache = null)
         {
             if (!File.Exists(item.FilePath)) return null;
+            if (cancellationToken.IsCancellationRequested)
+                return null;
 
             var dirName = Path.GetDirectoryName(destination);
             Directory.CreateDirectory(dirName);
 
             var buildFile = new BuildFile(item.FilePath, destination);
-
+            if (cancellationToken.IsCancellationRequested)
+                return null;
             var importer = item.Importer;
             var importedFile = importer?.Import(item.FilePath, importerContext);
             if (importedFile == null) return null;
-
+            if (cancellationToken.IsCancellationRequested)
+                return null;
             buildFile.Dependencies.AddRange(importerContext.Dependencies);
             cache?.AddDependencies(Path.GetDirectoryName(item.FilePath), importerContext.Dependencies);
 
@@ -236,7 +295,8 @@ namespace engenious.ContentTool.Builder
 
             var processedFile = processor?.Process(importedFile, item.FilePath, processorContext);
             if (processedFile == null) return null;
-
+            if (cancellationToken.IsCancellationRequested)
+                return null;
             buildFile.Dependencies.AddRange(processorContext.Dependencies);
             cache?.AddDependencies(Path.GetDirectoryName(item.FilePath), processorContext.Dependencies);
 
@@ -265,7 +325,6 @@ namespace engenious.ContentTool.Builder
                 writer.WriteObject(processedFile, typeWriter);
             }
 
-
             cache?.AddFile(item.FilePath, buildFile);
             buildFile.RefreshModifiedTime();
 
@@ -287,6 +346,13 @@ namespace engenious.ContentTool.Builder
             Abort = 4,
             Finished = 8,
             Built = 16
+        }
+
+        public void Dispose()
+        {
+            _buildThreadCancellation.Cancel();
+            _startBuild.Set();
+            _buildThread.Join();
         }
     }
 }
